server/requirements.txt
Django
django_rest_framework
django-cors-headers
djangorestframework-simplejwt
pyotp
qrcode
pyopenssl
cryptography
pillow